requirements.txt
amqp==2.6.1
asgiref==3.4.1
billiard==3.5.0.5
celery==4.1.1
Django==1.11.6
django-cors-headers==2.1.0
djangorestframework==3.6.3
importlib-metadata==4.8.3
kombu==4.2.0
psycopg2==2.7.3
pytz==2022.1
redis==2.10.6
sqlparse==0.4.2
typing-extensions==4.1.1
vine==1.3.0
zipp==3.6.0
